tests: Move stream tests into own file
